Fix for "Cannot read property 'length' of null"
For phrases that contained characters like " some text . . ." , which caused the match in line 43 to return null, an error was getting generated when attempting to read the length of null in the line 70. This proposed fix does resolve the issue, but there may be a better way by modifying the regex above to keep "null" from getting returned.